Improved data for lists and album pages
List tags are now properly attached to albums in lists for the `list` page, including favorites lists. Tag-search calls were also not getting tags and this was fixed as well. The get_album controller now pulls in Favorites and Connections as well to provide more data in a single call.